Using the script tag to store and load json information in single page applications
This code shows how to store a json config object in a script tag and load it into a Object, modify it, save it back to a s script tag and then load it back into an object. In reality the script tag is just good for initial delivery from Node or PHP generated pages. What the code accomplishes is not putting the JSON days in the actual JS code.